requirements.txt

azureml-core==1.56.0
azureml-dataset-runtime==1.56.0
azureml-defaults==1.56.0
azure-ml==0.0.1
azure-ml-component==0.9.18.post2
azureml-mlflow==1.56.0
azureml-contrib-services==1.56.0
azureml-contrib-services==1.56.0
azureml-automl-common-tools==1.56.0
torch-tb-profiler==0.4.3
azureml-inference-server-http~=1.0.0
inference-schema==1.7.2
MarkupSafe==2.1.5
regex
pybind11
bitsandbytes==0.43.1
git+https://github.com/huggingface/transformers.git
peft==0.11.1
accelerate==0.30.1
datasets
scipy
azure-identity
azure-ai-ml